Assigning List_find C254 compartment ID
Actual Error: Mirage_runtime__set_level(C15) calls Stdlib__list_find(C254) which in turn raises an exception and calls caml_raise_exn, which
in turn jumps directly to the corresponding exception handler which is registered in Mirage_runtime__set_level. By directly jumping to this, the CSR_UNOCROSSCOMP
is not reset, as it does the exception path does not follow the normal return path. Now while Mirage_runtime__set_level is executing the CSR_UNOCROSSCOMP is still 1,
and when it invokes Logs__set_level_inner, then cross-compartment call not allowed.
Solution:
1). Support cross-compartment exceptions
2). For now place Stdlib__list_find in C255. (Choosing this for now)