Fixed operator names in cap_tee files
The operator's function names generated as part of asm files/dump files is not correct. For eg: camlStdlib__@_1167 is replaced by camlStdlib__.5e.5e_201386. So the compartment IDs are not updated properly leading to error. For now doing this in manual way. Possible solution is to generate cap_tee_files based on compiler output instead of generating them based on the generated asm files